Fix appId and webui extraction for pyspark
