+++ livy/src/main/java/org/apache/zeppelin/livy/LivySparkInterpreter.java
@@ -44,4 +44,39 @@ public class LivySparkInterpreter extends BaseLivyInterprereter {
     return "spark";
   }
 
+  @Override
+  protected String extractAppId() throws LivyException {
+    return extractStatementResult(
+        interpret("sc.applicationId", null, false, false).message()
+            .get(0).getData());
+  }
+
+  @Override
+  protected String extractWebUIAddress() throws LivyException {
+    interpret(
+        "val webui=sc.getClass.getMethod(\"ui\").invoke(sc).asInstanceOf[Some[_]].get",
+        null, false, false);
+    return extractStatementResult(
+        interpret(
+            "webui.getClass.getMethod(\"appUIAddress\").invoke(webui)", null, false, false)
+            .message().get(0).getData());
+  }
+
+  /**
+   * Extract the eval result of spark shell, e.g. extract application_1473129941656_0048
+   * from following:
+   * res0: String = application_1473129941656_0048
+   *
+   * @param result
+   * @return
+   */
+  private String extractStatementResult(String result) {
+    int pos = -1;
+    if ((pos = result.indexOf("=")) >= 0) {
+      return result.substring(pos + 1).trim();
+    } else {
+      throw new RuntimeException("No result can be extracted from '" + result + "', " +
+          "something must be wrong");
+    }
+  }
 }
